pyproject.toml
[tool.poetry]
name = "azazel"
version = "0.2.0"
description = "A versatile Disord Bot designed to enhance server management."
authors = ["Daniel Lima"]
license = "apache-2.0"
readme = "README.md"
keywords = [
    "discord bot",
    "chatbot",
    "discord",
    "python",
    "moderation",
    "asyncio",
    "bot",
    "community management",
    "automation",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Communications :: Chat",
    "Topic :: Games/Entertainment",
    "Framework :: AsyncIO",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]

[tool.poetry.dependencies]
python = "^3.12"
discord-py = "^2.4.0"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
black = "^24.8.0"
pylint = "^3.3.0"
mypy = "^1.13.0"
codespell = "^2.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
